feat: when the payment entry is canceled, unlink the payment reference document
