+++ india_banking/india_banking/doc_events/payment_entry.py
@@ -1,0 +1,7 @@
+from india_banking.utils import get_payment_order_summary, unlink_bank_payment
+
+
+def on_cancel(doc, method=None):
+	payment_order_summary = get_payment_order_summary(doc.name)
+
+	unlink_bank_payment(payment_order_summary)
