+++ india_banking/utils.py
@@ -1,7 +1,13 @@
 import json
 
 import frappe
+from erpnext.accounts.doctype.accounting_dimension.accounting_dimension import (
+	get_accounting_dimensions,
+)
 from frappe import _
+from frappe.utils import cstr
+
+from india_banking.default import PAYMENT_SUMMARIES_FIELDS
 
 
 def get_bank_address_details(bank_account):
@@ -57,7 +63,11 @@ def get_party_field_name(party_type):
 
 def extract_error_message(response_json, show_message=False) -> str:
 	try:
-		response_json = json.loads(response_json) if isinstance(response_json, str) else response_json
+		response_json = (
+			json.loads(response_json)
+			if isinstance(response_json, str)
+			else response_json
+		)
 		failure_message = ""
 
 		server_message = response_json.get("_server_messages", "[]")
@@ -82,3 +92,48 @@ def extract_error_message(response_json, show_message=False) -> str:
 			title=_("Error: Could not process the response"),
 			msg=frappe.get_traceback(with_context=1),
 		)
+
+
+def unlink_bank_payment(payment_order_summary=None):
+	if not payment_order_summary:
+		return
+
+	payment_order = frappe.get_doc("Payment Order", payment_order_summary.parent)
+
+	summarise_field = PAYMENT_SUMMARIES_FIELDS.copy()
+	summarise_field.remove("payment_entry")
+	summarise_field.extend(get_accounting_dimensions())
+
+	if payment_order.summarise_payment_based_on == "Party":
+		summarise_field.remove("reference_name")
+
+	for reference in payment_order.references:
+		if all(
+			(
+				cstr(reference.get(field, ""))
+				== cstr(payment_order_summary.get(field, ""))
+				for field in summarise_field
+			)
+		):
+			frappe.db.set_value(
+				"Payment Request",
+				reference.payment_request,
+				{"reference_doctype": "", "reference_name": ""},
+			)
+			frappe.db.set_value(
+				"Payment Order Reference",
+				reference.name,
+				{"reference_doctype": "", "reference_name": ""},
+			)
+
+
+def get_payment_order_summary(payment_entry):
+	is_ammended = frappe.db.get_value("Payment Entry", payment_entry, "amended_from")
+	payment_entry = (
+		"-".join(payment_entry.split("-")[:-1]) if is_ammended else payment_entry
+	)
+	summary = frappe.db.get_value(
+		"Payment Order Summary", {"payment_entry": payment_entry}, "name"
+	)
+	if summary:
+		return frappe.get_doc("Payment Order Summary", summary)
